doc_requirements.txt
aiohttp
aiofiles
backoff
coreapi
django
djangorestframework
djangorestframework-queryfields
django-filter
django-import-export
drf-nested-routers
drf-yasg
dynaconf
plantuml
pyyaml
pygtrie
python-gnupg
rq
sphinx
sphinx-rtd-theme
sphinxcontrib-openapi
towncrier
